searchx/requirements.txt
Django==2.2
psycopg2==2.8.1
django-elasticsearch-dsl==7.1.0
djangorestframework==3.9.2
pyexcel==0.5.13
pyexcel-io==0.5.17
wagtail-django-recaptcha==1.0